# [Multithreading] 2 steps to maximize coalescing memory access